requirements.txt
### Core Dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# AI Assistant
anthropic>=0.39.0

# Data Processing
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization (for dashboard)
plotly>=5.15.0
streamlit>=1.25.0

# HTTP Client (for webhooks)
httpx>=0.25.0

# Caching (optional Redis support)
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# PDF Processing
pdfplumber>=0.10.0

# Web Templates
jinja2>=3.1.0
python-multipart>=0.0.6
